Move ColumnMetaDataLoader to binder module (#4739)
* revise #4577, cannot use rule in lexer

* refactor org.apache.shardingsphere.underlying.common.metadata.table

* refactor ColumnMetaDataLoader to static

* move isTableExist to ColumnMetaDataLoader
